Migrate render tests into chai-kefir style
This matches how this would actually be used by a consumer,
making the tests a useful form of documentation.
