+++ packages/brookjs/src/component/__tests__/fixtures/attachDetachReplace.hbs
