+++ packages/brookjs/src/component/__tests__/fixtures/attributeText.hbs
