+++ packages/brookjs/src/component/__tests__/fixtures/hideBlackboxed.hbs
