+++ packages/brookjs/src/component/__tests__/fixtures/rootBlackboxed.hbs
@@ -1,0 +1,1 @@
+<div {{container "root"}} {{blackbox "root"}}>{{text}}</div>
