+++ packages/brookjs/src/component/__tests__/fixtures/simpleUpdate.hbs
