+++ packages/brookjs/src/component/__tests__/fixtures/updateChild.hbs
