--- packages/brookjs/src/component/render/__tests__/fixtures/rootBlackboxed.hbs
@@ -1,1 +1,0 @@
-<div {{container "root"}}  {{blackbox "root"}}>{{text}}</div>
